requirements.txt
gcsfs
newsapi-python
pandas
json-merge-patch
json5
jsonschema
google-cloud-language
fsspec
